+++ RealTimeCollaborativeWhiteboard/Controllers/ViewNotesController.cs
@@ -1,9 +1,11 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using RealTimeCollaborativeWhiteboard.Data;
 using RealTimeCollaborativeWhiteboard.Models;
 using System.Composition;
+using System.Security.Claims;
 
 namespace RealTimeCollaborativeWhiteboard.Controllers
 {
@@ -18,6 +20,7 @@ namespace RealTimeCollaborativeWhiteboard.Controllers
         }
 
         [HttpGet]
+        [Authorize]
         public ActionResult Index()
         {
             var notes = _dbContext.Notes.ToList();
@@ -25,15 +28,21 @@ namespace RealTimeCollaborativeWhiteboard.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id != null)
             {
                 var note = await _dbContext.Notes.FirstOrDefaultAsync(p => p.NotesId == id);
-                if (note != null)
+                var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+                if (note.CurrUserID == userId)
                 {
-                    _dbContext.Notes.Remove(note);
-                    await _dbContext.SaveChangesAsync();
+                    if (note != null)
+                    {
+                        _dbContext.Notes.Remove(note);
+                        await _dbContext.SaveChangesAsync();
+                    }
                 }
             }
             return RedirectToAction("Index");
@@ -41,6 +50,7 @@ namespace RealTimeCollaborativeWhiteboard.Controllers
 
 
         [HttpGet]
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -54,30 +64,32 @@ namespace RealTimeCollaborativeWhiteboard.Controllers
                 return RedirectToAction("Index");
             }
 
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (note.CurrUserID != userId) {
+                return RedirectToAction("Index");
+            }
+
             return View(note);
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Edit(int id, Notes updateNote)
         {
             var note = await _dbContext.Notes.FirstOrDefaultAsync(p => p.NotesId == id);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             if (note != null)
             {
-                if (updateNote.Title != null)
-                {
-                    note.Title = updateNote.Title;
-                }
-                if (updateNote.Content != null)
-                {
-                    note.Content = updateNote.Content;
+                if(note.CurrUserID == userId) {
+                    if (updateNote.Title != null && updateNote.Content != null)
+                    {
+                        note.Title = updateNote.Title;
+                        note.Content = updateNote.Content;
+                    }
                 }
                 await _dbContext.SaveChangesAsync();
             }
             return RedirectToAction("Edit");
         }
-
-
-
-
     }
 }
